+++ bld/wv/h/madinter.h
@@ -40,7 +40,7 @@ extern void            GetMADTypeDefaultAt( address a, mad_type_kind mtk, mad_ty
 extern void            GetMADTypeDefault( mad_type_kind mtk, mad_type_info *mti );
 extern void            ReportMADFailure( mad_status ms );
 extern void            CheckMADChange( void );
-extern unsigned        GetMADNormalizedString( mad_string ms, char *buff, unsigned buff_len );
-extern unsigned        GetMADTypeNameForCmd( mad_type_handle th, char *buff, unsigned buff_len );
+extern size_t          GetMADNormalizedString( mad_string ms, char *buff, size_t buff_len );
+extern size_t          GetMADTypeNameForCmd( mad_type_handle th, char *buff, size_t buff_len );
 extern mad_type_handle FindMADTypeHandle( mad_type_kind tk, unsigned size );
 extern mad_handle      FindMAD( const char *name, unsigned len );
